refactor: break tools out from monitor into agent-utils
